fix: PCI_MAX_BUSES uint16_t for clang, Timer::ns 64-bit arithmetic, test infra rewrite
- src/kernel/arch/pci.hpp: uint8_t -> uint16_t (clang -Wconstant-conversion)
- src/kernel/arch/x86_64/timer.cpp: replace __int128 with pure 64-bit
  arithmetic (sec/rem) to avoid __udivti3 divide-overflow hang in bare metal
- src/kernel/arch/hal/timer.hpp: declare ns(), calibrate_tsc(), tsc_freq_hz_
- src/kernel/arch/x86_64/timer.cpp: PIT-based TSC calibration with retry loop
- src/kernel/test/test_watchdog.*: dual watchdog (TSC deadline + PIT ch1
  one-shot + PIT ch0 elapsed-time fallback), inline check at test boundaries
- src/kernel/task/scheduler.cpp: use ns-based watchdog deadline
- src/kernel/kernel.cpp: config-driven test classes, debug registry count
- src/lib/test.*: new S: format output (class suite::name n/m: PASS/FAIL),
  ResourceTracker leak detail on failure, serial FIFO drain, print_report(ns)
- src/kernel/memory/checked_ptr.hpp: allow zero-size null ptr operations
- Makefile: test-selftest, test-all-debug, test-all-release, test-class
  targets with expect regex summary validation + tee capture
- mk/rules.mk: include test-config.txt in initrd CPIO
- initrd/tests/test-config.txt: default 'safe' class
- src/kernel/test/test_config.*: config file parser
- PROMPT-dev.md, PROMPT-testdev.md, AGENTS.md: infrastructure docs

+++ src/kernel/arch/pci.hpp
@@ -34,7 +34,7 @@ constexpr uint16_t PCI_CONFIG_DATA = 0xCFC;
 /// Max devices per bus
 constexpr uint8_t  PCI_MAX_DEVICES   = 32;
 constexpr uint8_t  PCI_MAX_FUNCTIONS = 8;
-constexpr uint8_t  PCI_MAX_BUSES     = 256;
+constexpr uint16_t PCI_MAX_BUSES     = 256;
 
 /// Standard PCI config space register offsets
 enum PciRegister : uint8_t {
